requirements.txt
anthropic>=0.25.0
tavily-python>=0.3.0
firecrawl-py>=0.0.16
praw>=7.7.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
jinja2>=3.1.0
weasyprint>=60.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.9.0
